+++ CHANGELOG.md
@@ -7,6 +7,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - Remove `xhr2` dependency for NodeJS, use `fetch`
 - Add `mvs-stories` app included in the `molstar` NPM package
   - Use the app in the corresponding example
+- Interactions extension: remove `salt-bridge` interaction kind (since `ionic` is supported too)
 
 ## [v4.16.0] - 2025-05-20
 - Load potentially big text files as `StringLike` to bypass string size limit